+++ src/algorithm/reduce.rs
@@ -444,7 +444,9 @@ where
 }
 
 fn generic_reduce(f: Function, xs: Value, depth: usize, env: &mut Uiua) -> UiuaResult {
-    generic_reduce_impl(f, xs, depth, identity, env)
+    let val = generic_reduce_inner(f, xs, depth, identity, env)?;
+    env.push(val);
+    Ok(())
 }
 
 pub fn reduce_content(env: &mut Uiua) -> UiuaResult {
@@ -469,27 +471,8 @@ pub fn reduce_content(env: &mut Uiua) -> UiuaResult {
         env.push(acc);
         return Ok(());
     }
-    generic_reduce_impl(f, xs, 0, Value::unboxed, env)
-}
-
-fn generic_reduce_impl(
-    f: Function,
-    xs: Value,
-    depth: usize,
-    process: impl Fn(Value) -> Value + Copy,
-    env: &mut Uiua,
-) -> UiuaResult {
-    let sig = f.signature();
-    if let (0 | 1, 1) = (sig.args, sig.outputs) {
-        // Backwards compatibility for deprecated reduce behavior
-        for row in xs.into_rows() {
-            env.push(process(row));
-            env.call(f.clone())?;
-        }
-    } else {
-        let val = generic_reduce_inner(f, xs, depth, process, env)?;
-        env.push(val);
-    }
+    let val = generic_reduce_inner(f, xs, 0, Value::unboxed, env)?;
+    env.push(val);
     Ok(())
 }